app/requirements.txt
requests
httpx
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
asyncpg
tensorflow-cpu
pandas
numpy
prometheus-api-client
locust
joblib
scikit-learn
prometheus-client
python-multipart
